+++ packages/expo-router/src/__tests__/navigation.test.ios.tsx
@@ -868,42 +868,83 @@ it('can push the same route multiple times', () => {
   expect(screen.getByTestId('index')).toBeOnTheScreen();
 });
 
-it('can push relative links from index routes', async () => {
-  renderRouter({
-    _layout: () => <Slot />,
-    '(app)/index': () => <Text testID="one" />,
-    '(app)/test/_layout': () => <Stack />,
-    '(app)/test/index': () => <Text testID="two" />,
-    '(app)/test/bar': () => <Text testID="three" />,
+describe('relative urls', () => {
+  it('can push relative links from index routes', async () => {
+    renderRouter(
+      {
+        _layout: () => <Slot />,
+        '(app)/test/_layout': () => <Stack />,
+        '(app)/test/index': () => <Text testID="two" />,
+        '(app)/test/bar': () => <Text testID="three" />,
+      },
+      {
+        initialUrl: '/test',
+      }
+    );
+
+    expect(screen).toHavePathname('/test');
+    expect(screen.getByTestId('two')).toBeOnTheScreen();
+
+    act(() => router.push('./test/bar'));
+    expect(screen.getByTestId('three')).toBeOnTheScreen();
+    expect(screen).toHavePathname('/test/bar');
   });
 
-  expect(screen).toHavePathname('/');
-  expect(screen.getByTestId('one')).toBeOnTheScreen();
+  it('can push relative links relative to the directory', async () => {
+    renderRouter(
+      {
+        _layout: () => <Slot />,
+        '(app)/index': () => <Text testID="one" />,
+        '(app)/test/_layout': () => <Stack />,
+        '(app)/test/index': () => <Text testID="two" />,
+        '(app)/test/bar': () => <Text testID="three" />,
+      },
+      {
+        initialUrl: '/test',
+      }
+    );
 
-  act(() => router.push('./test'));
-  expect(screen).toHavePathname('/test');
-  expect(screen.getByTestId('two')).toBeOnTheScreen();
+    expect(screen).toHavePathname('/test');
+    expect(screen.getByTestId('two')).toBeOnTheScreen();
 
-  act(() => router.push('./bar'));
-  expect(screen.getByTestId('three')).toBeOnTheScreen();
-  expect(screen).toHavePathname('/test/bar');
-});
+    act(() => router.push('./bar', { relativeToDirectory: true }));
+    expect(screen.getByTestId('three')).toBeOnTheScreen();
+    expect(screen).toHavePathname('/test/bar');
+  });
 
-it('can push relative links from hoisted routes', () => {
-  renderRouter(
-    {
-      _layout: () => <Stack />,
-      'parent/index': () => <Link testID="link" href="./child" />,
-      'parent/child': () => <View testID="child" />,
-    },
-    {
-      initialUrl: '/parent',
-    }
-  );
+  it('can push relative links from hoisted routes', () => {
+    renderRouter(
+      {
+        _layout: () => <Stack />,
+        'parent/index': () => <Link testID="link" href="./parent/child" />,
+        'parent/child': () => <View testID="child" />,
+      },
+      {
+        initialUrl: '/parent',
+      }
+    );
 
-  expect(screen.getByTestId('link')).toBeOnTheScreen();
-  fireEvent(screen.getByTestId('link'), 'press');
-  expect(screen.getByTestId('child')).toBeOnTheScreen();
+    expect(screen.getByTestId('link')).toBeOnTheScreen();
+    fireEvent(screen.getByTestId('link'), 'press');
+    expect(screen.getByTestId('child')).toBeOnTheScreen();
+  });
+
+  it('can push relative links from hoisted routes relative to the directory', () => {
+    renderRouter(
+      {
+        _layout: () => <Stack />,
+        'parent/index': () => <Link testID="link" href="./child" relativeToDirectory />,
+        'parent/child': () => <View testID="child" />,
+      },
+      {
+        initialUrl: '/parent',
+      }
+    );
+
+    expect(screen.getByTestId('link')).toBeOnTheScreen();
+    fireEvent(screen.getByTestId('link'), 'press');
+    expect(screen.getByTestId('child')).toBeOnTheScreen();
+  });
 });
 
 it('can navigation to a relative route without losing path params', async () => {
@@ -1324,4 +1365,75 @@ describe('stack unwinding', () => {
 
     expect(router.canGoBack()).toBe(true); //
   });
+});
+
+it('should always prefer static routes over dynamic ones', async () => {
+  renderRouter(
+    {
+      // Uses Layouts at different levels to create different hoisting for each group
+      '(tabs)/nested/_layout': () => null,
+      '(tabs)/nested/index': () => null,
+      '(tabs)/nested/[fruit]': () => null,
+      '(tabs)/nested/orange': () => null,
+      '(stack)/_layout': () => null,
+      '(stack)/nested/banana': () => null,
+      '(stack)/nested/[fruit]': () => null,
+      'nested/grape': () => null,
+      'nested/[fruit]': () => null,
+      '[param]/melon': () => null,
+    },
+    {
+      initialUrl: '/(tabs)/nested/apple',
+    }
+  );
+
+  // We start in (tabs)
+  expect(screen).toHavePathname('/nested/apple');
+  expect(screen).toHaveSegments(['(tabs)', 'nested', '[fruit]']);
+
+  // Banana is more specific in (stack) so we move
+  act(() => router.push('/nested/banana'));
+  expect(screen).toHaveSegments(['(stack)', 'nested', 'banana']);
+
+  // Apple could be in either (tabs) or (stack) so we stay in the same group
+  act(() => router.push('/nested/apple'));
+  expect(screen).toHavePathname('/nested/apple');
+  expect(screen).toHaveSegments(['(stack)', 'nested', '[fruit]']);
+
+  // Orange is more specific in (tabs) so we move
+  act(() => router.push('/nested/orange'));
+  expect(screen).toHavePathname('/nested/orange');
+  expect(screen).toHaveSegments(['(tabs)', 'nested', 'orange']);
+
+  // Grape is more specific outside any group
+  act(() => router.push('/nested/grape'));
+  expect(screen).toHavePathname('/nested/grape');
+  expect(screen).toHaveSegments(['nested', 'grape']);
+
+  // This matches /(tabs)/nested/[fruit].
+  // We don't match:
+  // - nested/[fruit] because /(tabs)/nested/fruit is more specific
+  // - [param]/melon because segments are evaluated left-right. 'nested' is static and '[param]' is dynamic
+  act(() => router.push('/nested/melon'));
+  expect(screen).toHavePathname('/nested/melon');
+  expect(screen).toHaveSegments(['(tabs)', 'nested', '[fruit]']);
+});
+
+it('can push relative links that are relative to the directory', () => {
+  renderRouter(
+    {
+      '(stack)/_layout': () => <Stack />,
+      '(stack)/[fruit]/index': function Fruit() {
+        const { fruit } = useLocalSearchParams();
+        return <Text testID="fruit">{fruit}</Text>;
+      },
+    },
+    {
+      initialUrl: '/apple',
+    }
+  );
+
+  expect(screen.getByText('apple')).toBeOnTheScreen();
+  act(() => router.push('./banana'));
+  expect(screen.getByText('banana')).toBeOnTheScreen();
 });