+++ packages/expo-router/src/fork/getStateFromPath.ts
@@ -93,7 +93,7 @@ export function getUrlWithReactNavigationConcessions(
  * @param options Extra options to fine-tune how to parse the path.
  */
 export default function getStateFromPath<ParamList extends object>(
-  this: RouterStore | undefined,
+  this: RouterStore | undefined | void,
   path: string,
   options?: Options<ParamList>
 ): ResultState | undefined {
@@ -236,8 +236,30 @@ function sortConfigs(a: RouteConfig, b: RouteConfig, previousSegments: string[] 
     bParts.push('index');
   }
 
-  // When we navigate, we need to stay within groups as close as possible
-  // Hence, a route is sorted based upon is similiarity to the current state
+  const isAStaticRoute =
+    !a.hasChildren && // Layout configs will have children
+    !aParts.some(
+      (part) => part.startsWith(':') || part.startsWith('*') || part.includes('*not-found')
+    );
+  const isBStaticRoute =
+    !b.hasChildren &&
+    !bParts.some(
+      (part) => part.startsWith(':') || part.startsWith('*') || part.includes('*not-found')
+    );
+
+  /*
+   * Static routes should always be higher than dynamic routes.
+   * Layouts are excluded from this and are ranked lower than routes
+   */
+  if (isAStaticRoute && !isBStaticRoute) {
+    return -1;
+  } else if (!isAStaticRoute && isBStaticRoute) {
+    return 1;
+  }
+
+  /*
+   * If both are static/dynamic or a layout file, then we check group similarity
+   */
   const similarToPreviousA = previousSegments.filter((value, index) => {
     return value === a.expandedRouteNames[index] && value.startsWith('(') && value.endsWith(')');
   });
@@ -250,10 +272,13 @@ function sortConfigs(a: RouteConfig, b: RouteConfig, previousSegments: string[] 
     (similarToPreviousA.length > 0 || similarToPreviousB.length > 0) &&
     similarToPreviousA.length !== similarToPreviousB.length
   ) {
-    // They both match to some degree, so pick the one that matches more
+    // One matches more than the other, so pick the one that matches more
     return similarToPreviousB.length - similarToPreviousA.length;
   }
 
+  /*
+   * If there is not difference in similarity, then each non-group segment is compared against each other
+   */
   for (let i = 0; i < Math.max(aParts.length, bParts.length); i++) {
     // if b is longer, b get higher priority
     if (aParts[i] == null) {
@@ -316,12 +341,24 @@ function sortConfigs(a: RouteConfig, b: RouteConfig, previousSegments: string[] 
     }
   }
 
-  // Sort initial routes with a higher priority than routes which will push more screens
-  // this ensures shared routes go to the shortest path.
+  /*
+   * Both configs are identical in specificity and segments count/type
+   * Try and sort by initial instead.
+   *
+   * TODO: We don't differentiate between the default initialRoute and group specific default routes
+   *
+   * const unstable_settings = {
+   *   "group": {
+   *     initialRouteName: "article"
+   *  }
+   * }
+   *
+   * "article" will be ranked higher because its an initialRoute for a group - even if not your not currently in
+   * that group. The current work around is to ways provide initialRouteName for all groups
+   */
   if (a.isInitial && !b.isInitial) {
     return -1;
-  }
-  if (!a.isInitial && b.isInitial) {
+  } else if (!a.isInitial && b.isInitial) {
     return 1;
   }
 
@@ -671,9 +708,9 @@ const createConfigItem = (
     path,
     // The routeNames array is mutated, so copy it to keep the current state
     routeNames: [...routeNames],
-    expandedRouteNames: screen.includes('/')
-      ? [...routeNames.slice(0, -1), ...screen.split('/')]
-      : [...routeNames],
+    expandedRouteNames: routeNames.flatMap((name) => {
+      return name.split('/');
+    }),
     parse,
     userReadableName: [...routeNames.slice(0, -1), path || screen].join('/'),
     hasChildren: !!hasChildren,